3D rendering and depth testing
Rewrote all code for affine transformations, incl creating
view, model, and projection matrices, using glm's matrix
and vector types. The storage is row-major and right-handed
I THINK, but in case that's wrong, here's the order of ops:

  1. When multiplying a vector by a matrix, the vector goes
     to the right of the matrix.

  2. Matrices do NOT need transposed before being passed to
     the shader.

  3. When applying a transform thru matrix multiplication,
     The pre-existing matrix goes on the left, and the new
     Transformation goes on the right. All matrix ops such
     as calculating the MVP matrix should sequentially
     proceed from left to right.

  4. Normals are inverted so that x is positive to the
     right, y is positive upward, and z is positive in
     the reverse of the direction that the camera isf
     facing (so backward).

  5. Projection is passed HALF of the vertical fov (the
     angle from the xz-plane to the top of the frustum) so
     that, when coupled with the aspect ratio, we can com-
     pute the top, bottom, left and right of the viewing
     plane and use these for the frustum. Another quirk of
     projection is that the near and far planes are
     in reverse-z, so to achieve the best floating-point
     distribution when depth sampling, we pass in 1.0f as
     near and 0.0f as far, projecting everything onto
     the plane z = -1.

Created a Context initializer that includes a render pipeline
for 3D rendering, Buffers for scene vertices, indices, normals
and draw commands, handles for the SDL_Window and SDL_GPUDevice
that are being used by the application, a depth texture, and a
Color Texture that is as-yet unused but may be sampled for use
in an effects pipeline.

Defined one Vertex type to pass into the buffer
(the PositionAndColorVertex type) that stores position coords
as vectors of 3 32-bit floating point numbers, color values
as 4 unsigned bytes. Normals are also stored as 3 32-bit
floating point values.

Defined naive wrappers for the state of a 3D object that
include that object's vertex data and the indices for its
instances, its normal data in object-space, and a 4x4
matrix that represents the transform to move each vertex
from object-space to world-space. Draw commands must be created
and loaded to the Draw Buffer when the scene is initialized,
this is not handled by the RenderableObject struct.

Defined simple wrapper for the state of a 3D camera including
field-of-view, aspect ratio, near plane, far plane, and transforms
from world-space into view-space
and from view-space into clip-space. Camera also memoizes its
own coordinates and those of its target for use in updating the
view-space transform and for easy transformations——as long as you
have position data for two vertices, you can generate a view-space
transform.

Created simple tracking of keypresses to allow for interactive
transformations and some simple event-handling and update functions
for a test scene that involves rotating a cube.

Also included several shaders that were written for testing that
consume Position-only vertices, Position-and-Color vertices, and
several fragment shaders that interpolate vertex colors, output
only white, and write fragment w-coords to the depth buffer.